fix: update systems to not depend on Environment.TickCount

+++ Client/Components/Movement/CharacterStateComponent.cs
@@ -3,5 +3,12 @@ namespace CryBits.Client.Components.Movement;
 internal struct CharacterStateComponent
 {
     public bool IsAttacking;
-    public int AttackTimer;
+
+    /// <summary>
+    /// Seconds remaining in the current attack cooldown.
+    /// Written as <c>AttackSpeed / 1000f</c> when an attack starts;
+    /// ticked down by dt in <see cref="Systems.Movement.CharacterAnimationControllerSystem"/>.
+    /// <see cref="IsAttacking"/> is cleared when this reaches zero.
+    /// </summary>
+    public float AttackTimer;
 }
